merge main to develop

+++ src/client/interfaces/psyneulinkHandler.js
@@ -1,24 +1,30 @@
 const os = require('os');
 const { resolve } = require("path");
+const sleep = require("./utils").sleep;
 const logOutput = require("./utils").logOutput;
 const killProcess = require("./utils").killProcess;
 const spawnCommand = require("./utils").spawnCommand;
 const executeCommand = require("./utils").executeCommand;
+const executeSyncCommand = require("./utils").executeSyncCommand;
+const parseArguments = require("./utils").parseArguments;
+
+const environments = require("../../nodeConstants").environments;
 
 const psyneulinkHandlerFactory = (function(){
     function PsyneulinkHandler() {
-        this.condaEnv = null;
+        this.condaEnv = executeSyncCommand('conda info').split("\n").filter((item) => { return item.includes("active environment") })[0].split(":")[1].trim();
+        // this.serverProc = null;
         this.psyneulinkInstalled = false;
-        this.serverProc = null;
+        this.environment = parseArguments(process.argv)['--mode'] || environments.PROD;
 
-        this.isPsyneulinkInstalled = async () => {
-            const pipPsyneuLink = await this.runCommand("pip show psyneulink");
+        this.isPsyneulinkInstalled = () => {
+            const pipPsyneuLink = this.runSyncCommand("pip show psyneulink");
             this.psyneulinkInstalled = pipPsyneuLink.includes("Name: psyneulink");
             return this.psyneulinkInstalled;
         }
 
         this.getCondaEnvs = async () => {
-            const condaEnvList = await this.runCommand("conda env list");
+            const condaEnvList = await this.runSyncCommand("conda env list");
             const allEnvs = condaEnvList.split("\n").slice(2, -1).filter(
                 (item) => { return item !== "" }).map(
                     (item) => { return item.split(" ")[0] });
@@ -29,6 +35,12 @@ const psyneulinkHandlerFactory = (function(){
             return this.condaEnv;
         }
 
+        this.getRunningEnv = () => {
+            const results = executeSyncCommand("conda info");
+            const env = results.split("\n").filter((item) => { return item.includes("active environment") })[0].split(":")[1].trim();
+            return env;
+        }
+
         this.runCommand = async (command) => {
             if (this.condaEnv) {
                 command = `conda run -n ${this.condaEnv} ${command}`;
@@ -37,9 +49,23 @@ const psyneulinkHandlerFactory = (function(){
             return result;
         }
 
-        this.installViewerDependencies = async () => {
-            let result = await this.runCommand("pip install -r requirements.txt");
-            logOutput(Date.now() + " INFO: " + result + "\n", true);
+        this.runSyncCommand = (command) => {
+            if (this.condaEnv) {
+                command = `conda run -n ${this.condaEnv} ${command}`;
+            }
+            const result = executeSyncCommand(command);
+            return result;
+        }
+
+        this.installViewerDependencies = () => {
+            try {
+                let result = this.runSyncCommand("pip install -r requirements.txt");
+                logOutput(Date.now() + " INFO: " + result + "\n", true);
+                return true;
+            } catch (error) {
+                logOutput(Date.now() + " ERROR: " + error + "\n", true);
+                return false;
+            }
         }
 
         this.installPsyneulink = async () => {
@@ -58,36 +84,73 @@ const psyneulinkHandlerFactory = (function(){
             if (condaEnvList.includes(condaEnv)) {
                 this.condaEnv = condaEnv;
             } else {
-                throw new Error(`Conda environment ${condaEnv} not found.`);
+                this.condaEnv = "";
+                console.error("The conda environment " + condaEnv + " does not exist.");
             }
         }
 
-        this.runServer = () => {
-            if (this.serverProc) {
-                return;
-            }
-            const pythonServer = "python " + resolve(__dirname, "../../server/rpc_server.py");
-            this.serverProc =  spawnCommand(pythonServer, [], { condaEnv: this.condaEnv, isWin: os.platform() === "win32" });
+        this.runServer = async () => {
+            try {
+                // TODO - remove this when we have a proper server
+                if (this.environment === environments.DEV) {
+                    this.serverProc = 'DEVELOPMENT MODE';
+                    logOutput(Date.now() + " START: Starting Python RPC server IN DEVELOPMENT MODE\n", true);
+                    return true;
+                }
 
-            logOutput(Date.now() + " START: Starting Python RPC server \n", true);
-            
-            this.serverProc.on('error', function (err) {
-                logOutput(Date.now() + " ERROR: " + err + "\n", true);
-            });
-            this.serverProc.stdout.setEncoding('utf8');
-            this.serverProc.stdout.on('data', function (data) {
-                logOutput(Date.now() + " INFO: " + data + "\n", true);
-            });
-            this.serverProc.stderr.setEncoding('utf8');
-            this.serverProc.stderr.on('data', function (data) {
-                logOutput(Date.now() + " ERROR: " + data + "\n", true);
-            });
+                if (this.serverProc) {
+                    return;
+                }
+                const pythonServer = "python " + resolve(__dirname, "../../server/rpc_server.py");
+                this.serverProc =  spawnCommand(pythonServer, [], { condaEnv: this.condaEnv, isWin: os.platform() === "win32" });
+                let serverStarted = false;
+                while (!serverStarted) {
+                    const buffer = this.serverProc.stdout.read();
+                    if (buffer !== null && buffer.toString().includes("### PsyNeuLinkViewer Server UP ###")) {
+                        serverStarted = true;
+                    }
+                    await sleep(2000);
+                }
+                logOutput(Date.now() + " START: Starting Python RPC server \n", true);
+
+                this.serverProc.on('error', function (err) {
+                    logOutput(Date.now() + " ERROR: " + err + "\n", true);
+                });
+
+                this.serverProc.stdout.on('data', function (data) {
+                    logOutput(Date.now() + " INFO: " + data + "\n", true);
+                    if (data.toString().includes("### PsyNeuLinkViewer Server UP ###")) {
+                        serverStarted = true;
+                    }
+                });
+
+                this.serverProc.stderr.on('data', function (data) {
+                    logOutput(Date.now() + " ERROR: " + data + "\n", true);
+                });
+                return true;
+            } catch (error) {
+                logOutput(Date.now() + " ERROR: " + error + "\n", true);
+                return false;
+            }
         }
 
         this.stopServer = async () => {
-            if (this.serverProc) {
-                killProcess(this.serverProc.pid);
-                this.serverProc = null;
+            try {
+                if (this.environment === environments.DEV) {
+                    this.serverProc = 'DEVELOPMENT MODE';
+                    logOutput(Date.now() + " STOP: Simulation of the development server STOPPED\n", true);
+                    return true;
+                }
+
+                if (this.serverProc !== null && this.serverProc !== undefined) {
+                    killProcess(this.serverProc.pid);
+                    logOutput(Date.now() + " STOP: Server STOPPED with pid " + this.serverProc.pid + "\n", true);
+                    this.serverProc = null;
+                }
+                return true;
+            } catch (error) {
+                logOutput(Date.now() + " ERROR: " + error + "\n", true);
+                return false
             }
         }
     }
@@ -95,7 +158,7 @@ const psyneulinkHandlerFactory = (function(){
     var instance;
     return {
         getInstance: function(){
-            if (instance == null) {
+            if (instance == null || instance === undefined) {
                 instance = new PsyneulinkHandler();
                 instance.constructor = null;
             }
